POSXML review
This is the final commit of this review, totalling 133/133 reviews completed.

Reviewed and fixed EN content and replaced POSXML HTML with public Gists.

Removed file for command <input.getvalue> as it is not required.

Performed a complete translation from PT-BR to EN for the tables listed in <emv.getinfo> and <emv.adddate>